+++ DoorDeliveryApplication/src/UI/Customer/SearchCustomerResultsJPanel.java
@@ -246,7 +246,7 @@ public class SearchCustomerResultsJPanel extends javax.swing.JPanel {
         }
 
         String name = txtName.getText();
-        int phone = Integer.parseInt(txtPhone.getText());
+        long phone = Long.parseLong(txtPhone.getText());
         int zipcode = Integer.parseInt(txtZipcode.getText());
         String streetaddress = txtAddress.getText();
         String emailaddress = txtEmail.getText();
